move some tests to RestaurantCard.v1
