+++ packages/components/molecules/f-restaurant-card/src/components/_tests/RestaurantCard.v1.test.js
@@ -40,4 +40,32 @@ describe('RestaurantCard.v1', () => {
             expect(wrapper.find('[error-boundary][tier="3"]>[badges-slot]').exists()).toBe(true);
         });
     });
+
+    describe('Restaurant cuisines', () => {
+        it('should not be shown if there is no cuisines data', () => {
+            // arrange
+            const propsData = {
+                cuisines: []
+            };
+
+            // act
+            const wrapper = mount(RestaurantCardV1, { propsData });
+
+            // assert
+            expect(wrapper.find('[data-test-id="restaurant-cuisines"]').exists()).toBe(false);
+        });
+
+        it('should be shown if there is cuisines data', () => {
+            // arrange
+            const propsData = {
+                cuisines: ['Mexican', 'Burgers', 'Chinese']
+            };
+
+            // act
+            const wrapper = mount(RestaurantCardV1, { propsData });
+
+            // assert
+            expect(wrapper.find('[data-test-id="restaurant-cuisines"]').exists()).toBe(true);
+        });
+    });
 });
